apps/api/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[project]
name = "atlas-api"
version = "0.0.1"
description = "Atlas Cloud Cost Optimization API"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    # Web framework
    "fastapi>=0.111.0",
    "uvicorn[standard]>=0.29.0",

    # Validation / serialization
    "pydantic>=2.7.0",
    "pydantic-settings>=2.2.1",

    # Database
    "sqlalchemy[asyncio]>=2.0.29",
    "asyncpg>=0.29.0",
    "alembic>=1.13.1",

    # Redis
    "redis[hiredis]>=5.0.4",

    # S3 / MinIO
    "aiobotocore>=2.13.0",

    # Utilities
    "python-multipart>=0.0.9",
    "httpx>=0.27.0",
    "structlog>=24.1.0",
    "python-dotenv>=1.0.1",
    "tenacity>=8.2.3",
    "prometheus-client>=0.20.0",
    "opentelemetry-api>=1.24.0",
    "opentelemetry-sdk>=1.24.0",
    "opentelemetry-exporter-otlp-proto-http>=1.24.0",
    "opentelemetry-instrumentation-fastapi>=0.45b0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.2.0",
    "pytest-asyncio>=0.23.6",
    "pytest-cov>=5.0.0",
    "httpx>=0.27.0",
    "anyio>=4.3.0",
    "ruff>=0.6.0",
]

[tool.hatch.envs.default]
features = ["dev"]

# ─── Ruff ─────────────────────────────────────────────────────────────────────
[tool.ruff]
target-version = "py311"
line-length = 100
src = ["app"]

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "N",   # pep8-naming
    "UP",  # pyupgrade
    "B",   # flake8-bugbear
    "SIM", # flake8-simplify
    "ANN", # flake8-annotations
    "RUF", # ruff-specific
]
ignore = [
    "ANN101", # self type annotation
    "ANN102", # cls type annotation
    "ANN401", # Dynamically typed Any
]

[tool.ruff.lint.isort]
known-first-party = ["app"]
force-single-line = true

# ─── Black ────────────────────────────────────────────────────────────────────
[tool.black]
target-version = ["py311"]
line-length = 100

# ─── Mypy ─────────────────────────────────────────────────────────────────────
[tool.mypy]
python_version = "3.11"
strict = true
plugins = ["pydantic.mypy"]
ignore_missing_imports = true
disallow_untyped_defs = true
disallow_any_unimported = false

[[tool.mypy.overrides]]
module = ["tests.*"]
disallow_untyped_defs = false

# ─── Pytest ───────────────────────────────────────────────────────────────────
[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
addopts = "-v --tb=short --cov=app --cov-report=term-missing"

# ─── Coverage ─────────────────────────────────────────────────────────────────
[tool.coverage.run]
source = ["app"]
omit = ["tests/*", "alembic/*"]
